Add files via upload
Career portfolios help document education, work samples and skills. People use career portfolios to apply for jobs, apply to college or training programs. They are more in-depth than a resume, which is used to summarize the above in one or two pages. Career portfolios serve as proof of one's skills, abilities, and potential in the future. Career portfolios are becoming common in high schools, college, and workforce development.

Career portfolios help with a job or acceptance into higher education institutes. A career portfolio should be personal and contain critical information. Items that should be included include (but are not limited to) personal information, evaluations, sample work, a business portrait, and awards and acknowledgments. Career portfolios are often kept in a simple three-ring binder or online as an electronic portfolio and updated often. A career portfolio is used as a marketing tool in selling oneself for personal advancement. In some industries, employers or admission offices commonly request a career portfolio, so it is a wise idea to have an updated one on hand.